Split manual assign tests
